tangram_v3_requirements.txt
aiosignal==1.3.1
anndata==0.9.2
array-record==0.4.0
asciitree==0.3.3
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.1.1
cycler==0.12.1
dm-tree==0.1.8
docrep==0.3.2
etils==1.3.0
fasteners==0.19
filelock==3.14.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
gdown==5.2.0
get-annotations==0.1.2
gin-config==0.5.0
google-auth==2.30.0
google-pasta==0.2.0
googleapis-common-protos==1.63.2
h5py==3.11.0
idna==3.7
igraph==0.11.5
imageio==2.34.1
importlib_metadata==8.0.0
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.38.1
locket==1.0.0
loguru==0.7.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
natsort==8.4.0
networkx==3.1
numba==0.55.2
numcodecs==0.12.1
numpy==1.22.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
partd==1.4.1
patsy==0.5.6
pillow==10.3.0
promise==2.3
protobuf==3.20.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pynndescent==0.5.12
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scanpy==1.9.8
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
session-info==1.0.0
six==1.16.0
slicerator==1.1.0
statsmodels==0.14.1
stdlib-list==0.10.0
sympy==1.12.1
tangram-sc==0.4.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
tensorflow-probability==0.21.0
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2023.7.10
toolz==0.12.1
torch==2.3.0
tqdm==4.66.4
triton==2.3.0
typeguard==4.3.0
typing_extensions==4.12.1
tzdata==2024.1
umap-learn==0.5.6
urllib3==2.2.1
validators==0.28.3
wrapt==1.16.0
yarl==1.9.4
zarr==2.16.1
zipp==3.19.1
